Infinity Tdrop Approximation
Tdrop z index is not exactly the same as the Zeta Solver points, so the closest point calculation has changed.
First the solver checks if a point is close in the z, then it finds the XY distance.
In a sense projecting the Tdrop line onto the Zeta Solver points before calculating the distance.